backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
cryptography==44.0.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
google-auth==2.36.0
email-validator==2.2.0
